Add S3 bucket ACL and content type constants
Updated `MinioResourceBuilderExtensions` to include a
`CannedACL` property for public access when creating
buckets. Added a project reference to `HexaContent.Core.Utils`
in the project file. Enhanced the `Worker` class to upload
an HTML file to each bucket with the appropriate content type
and ACL. Introduced a new `ContentTypes` class for common
content type constants.