Add header management and update Varnish TTL settings
In the `ConfigExtensions.cs` file, a list of headers to remove
is introduced, along with a loop to remove these headers from
the response. A new response header "server" with the value
"Difoosion Proxy" is also added.

In the `VarnishResourceBuilderExtensions.cs` file, the
`beresp.ttl` is set to 1 minute, and the previous setting for
`beresp.keep` is removed. The `beresp.do_esi` setting remains
unchanged.